refactor(proxy): extract index scan handlers
Move index cursor entry points and index scan buffer reset into index_scan.cc.
Register index_scan.cc in the proxy build.
